third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/access-reporting-openee-rp-ro.https-expected.txt

This is a testharness.js-based test.
[FAIL] access-reporting-openee-rp-ro
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
